+++ CHANGELOG.md
@@ -65,8 +65,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Gem version badge to README
 - Usage documentation
 
-[Unreleased]: https://github.com/chris-allen/dream-ops/compare/v0.5.0...HEAD
-[0.4.2]: https://github.com/chris-allen/dream-ops/compare/v0.4.2...v0.5.0
+[Unreleased]: https://github.com/chris-allen/dream-ops/compare/v0.6.1...HEAD
+[0.6.1]: https://github.com/chris-allen/dream-ops/compare/v0.6.0...v0.6.1
+[0.6.0]: https://github.com/chris-allen/dream-ops/compare/v0.5.0...v0.6.0
+[0.5.0]: https://github.com/chris-allen/dream-ops/compare/v0.4.2...v0.5.0
 [0.4.2]: https://github.com/chris-allen/dream-ops/compare/v0.4.1...v0.4.2
 [0.4.1]: https://github.com/chris-allen/dream-ops/compare/v0.4.0...v0.4.1
 [0.4.0]: https://github.com/chris-allen/dream-ops/compare/v0.3.0...v0.4.0